--- src/main/java/com/swjungle/board/post/dto/PostListResponse.java
@@ -1,4 +1,0 @@
-package com.swjungle.board.post.dto;
-
-public class PostListResponse {
-}
