+++ src/main/java/com/swjungle/board/post/dto/request/CreatePostRequest.java
@@ -1,3 +1,3 @@
-package com.swjungle.board.post.dto;
+package com.swjungle.board.post.dto.request;
 
 public record CreatePostRequest(String title, String content, String link, String category, int score, String author, String password) {}
