+++ src/main/java/com/swjungle/board/post/dto/request/DeletePostRequest.java
@@ -1,4 +1,4 @@
-package com.swjungle.board.post.dto;
+package com.swjungle.board.post.dto.request;
 
 public record DeletePostRequest(String password) {
 }
