+++ src/main/java/com/swjungle/board/post/dto/request/UpdatePostRequest.java
@@ -1,4 +1,4 @@
-package com.swjungle.board.post.dto;
+package com.swjungle.board.post.dto.request;
 
 public record UpdatePostRequest(String title, String content, String link, String category, int score, String author, String password) {
 }
