+++ src/main/java/com/swjungle/board/post/dto/response/PostListResponse.java
@@ -1,0 +1,4 @@
+package com.swjungle.board.post.dto.response;
+
+public class PostListResponse {
+}
