+++ src/main/java/com/swjungle/board/post/dto/response/PostWrapperResponse.java
@@ -1,4 +1,4 @@
-package com.swjungle.board.post.dto;
+package com.swjungle.board.post.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
