+++ src/main/java/com/swjungle/board/post/service/PostService.java
@@ -1,19 +1,15 @@
 package com.swjungle.board.post.service;
 
 
-import com.swjungle.board.common.dto.EnvelopeResponse;
 import com.swjungle.board.common.dto.MessageResponse;
-import com.swjungle.board.post.dto.CreatePostRequest;
-import com.swjungle.board.post.dto.PostResponse;
-import com.swjungle.board.post.dto.PostWrapperListResponse;
-import com.swjungle.board.post.dto.UpdatePostRequest;
+import com.swjungle.board.post.dto.request.CreatePostRequest;
+import com.swjungle.board.post.dto.response.PostResponse;
+import com.swjungle.board.post.dto.request.UpdatePostRequest;
 import com.swjungle.board.post.entity.Post;
 import com.swjungle.board.post.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
