+++ src/test/java/com/swjungle/board/post/PostControllerTests.java
@@ -3,9 +3,9 @@ package com.swjungle.board.post;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.swjungle.board.common.dto.MessageResponse;
 import com.swjungle.board.post.controller.PostController;
-import com.swjungle.board.post.dto.CreatePostRequest;
-import com.swjungle.board.post.dto.PostResponse;
-import com.swjungle.board.post.dto.UpdatePostRequest;
+import com.swjungle.board.post.dto.request.CreatePostRequest;
+import com.swjungle.board.post.dto.response.PostResponse;
+import com.swjungle.board.post.dto.request.UpdatePostRequest;
 import com.swjungle.board.post.entity.Post;
 import com.swjungle.board.post.service.PostService;
 import org.hamcrest.Matchers;
