+++ src/test/java/com/swjungle/board/post/PostServiceTests.java
@@ -1,9 +1,9 @@
 package com.swjungle.board.post;
 
 import com.swjungle.board.common.dto.MessageResponse;
-import com.swjungle.board.post.dto.CreatePostRequest;
-import com.swjungle.board.post.dto.PostResponse;
-import com.swjungle.board.post.dto.UpdatePostRequest;
+import com.swjungle.board.post.dto.request.CreatePostRequest;
+import com.swjungle.board.post.dto.response.PostResponse;
+import com.swjungle.board.post.dto.request.UpdatePostRequest;
 import com.swjungle.board.post.entity.Post;
 import com.swjungle.board.post.repository.PostRepository;
 import com.swjungle.board.post.service.PostService;
